refactor: seperate bot and worker instances
