+++ submissions/chimoney-discord-bot/bot/deploy-commands.js
