--- submissions/chimoney-discord-bot/dockerfile
@@ -1,13 +1,0 @@
-FROM node:16-alpine
-
-WORKDIR /usr/src/app
-
-COPY package*.json ./
-
-RUN npm install
-
-COPY . .
-
-EXPOSE 8080
-
-CMD ["npm", "start"]
